c8d/commit-builder: Extract common code
Extract duplicated image creation code to a function.
